refactor: common case fields schema
